+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
+- Add `_from` variants of functions to support in-process threaded tests for
+  rustup.
+
 ## [0.5.3] - 2020-01-07
 
 Use Rust 1.36.0 as minimum Rust version.